+++ iothub/device/src/Authentication/DeviceAuthenticationWithTokenRefresh.cs
@@ -97,8 +97,8 @@ namespace Microsoft.Azure.Devices.Client
         /// Populates a supplied instance based on the properties of the current instance.
         /// </summary>
         /// <param name="iotHubConnectionStringBuilder">Instance to populate.</param>
-        /// <returns>The populated <see cref="IotHubConnectionStringBuilder"/> instance.</returns>
-        public override IotHubConnectionStringBuilder Populate(IotHubConnectionStringBuilder iotHubConnectionStringBuilder)
+        /// <returns>The populated <see cref="IotHubConnectionCredentials"/> instance.</returns>
+        public override IotHubConnectionCredentials Populate(IotHubConnectionCredentials iotHubConnectionStringBuilder)
         {
             iotHubConnectionStringBuilder = base.Populate(iotHubConnectionStringBuilder);
             iotHubConnectionStringBuilder.DeviceId = DeviceId;
