+++ iothub/device/src/Authentication/IAuthenticationMethod.cs
@@ -9,10 +9,10 @@ namespace Microsoft.Azure.Devices.Client
     public interface IAuthenticationMethod
     {
         /// <summary>
-        /// Populates an <see cref="IotHubConnectionStringBuilder"/> instance based on the properties of the current instance.
+        /// Populates an <see cref="IotHubConnectionCredentials"/> instance based on the properties of the current instance.
         /// </summary>
         /// <param name="iotHubConnectionStringBuilder">Instance to populate.</param>
-        /// <returns>The populated <see cref="IotHubConnectionStringBuilder"/> instance.</returns>
-        IotHubConnectionStringBuilder Populate(IotHubConnectionStringBuilder iotHubConnectionStringBuilder);
+        /// <returns>The populated <see cref="IotHubConnectionCredentials"/> instance.</returns>
+        IotHubConnectionCredentials Populate(IotHubConnectionCredentials iotHubConnectionStringBuilder);
     }
 }
