+++ iothub/device/tests/Authentication/AuthenticationWithTokenRefreshTests.cs
@@ -74,7 +74,7 @@ namespace Microsoft.Azure.Devices.Client.Test
         public async Task AuthenticationWithTokenRefresh_Populate_DefaultParameters_Ok()
         {
             var refresher = new TestImplementation();
-            var csBuilder = new IotHubConnectionStringBuilder(
+            var csBuilder = new IotHubConnectionCredentials(
                 new ModuleAuthenticationWithRegistrySymmetricKey("deviceId", "moduleid", TestSharedAccessKey),
                 TestIotHubName);
