+++ iothub/device/tests/Transport/Amqp/MockableAmqpUnit.cs
@@ -18,9 +18,9 @@ namespace Microsoft.Azure.Devices.Client.Test.Transport
 
         public MockableAmqpUnit()
             : this(
-                  new ClientConfiguration(new IotHubConnectionStringBuilder(AmqpTransportHandlerTests.TestConnectionString), new IotHubClientOptions(s_transportSettings)),
+                  new ClientConfiguration(new IotHubConnectionCredentials(AmqpTransportHandlerTests.TestConnectionString), new IotHubClientOptions(s_transportSettings)),
                   new AmqpConnectionHolder(
-                      new ClientConfiguration(new IotHubConnectionStringBuilder(AmqpTransportHandlerTests.TestConnectionString), new IotHubClientOptions(s_transportSettings))))
+                      new ClientConfiguration(new IotHubConnectionCredentials(AmqpTransportHandlerTests.TestConnectionString), new IotHubClientOptions(s_transportSettings))))
         {
         }
