+++ iothub/device/tests/Transport/Mqtt/MqttTransportHandlerTests.cs
@@ -556,7 +556,7 @@ namespace Microsoft.Azure.Devices.Client.Test.Transport
             return new MqttTransportHandler(
                 new PipelineContext
                 {
-                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionStringBuilder(DummyConnectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
+                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionCredentials(DummyConnectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
                 },
                 new IotHubClientMqttSettings());
         }
@@ -592,7 +592,7 @@ namespace Microsoft.Azure.Devices.Client.Test.Transport
             transport = new MqttTransportHandler(
                 new PipelineContext
                 {
-                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionStringBuilder(connectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
+                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionCredentials(connectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
                 },
                 new IotHubClientMqttSettings(),
                 factory);
@@ -606,7 +606,7 @@ namespace Microsoft.Azure.Devices.Client.Test.Transport
             return new MqttTransportHandler(
                 new PipelineContext
                 {
-                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionStringBuilder(connectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
+                    ClientConfiguration = new ClientConfiguration(new IotHubConnectionCredentials(connectionString), new IotHubClientOptions(new IotHubClientMqttSettings())),
                 },
                 new IotHubClientMqttSettings(),
                 null);
